requirements.txt
django>=1.4.2
djangorestframework+=2.1.1
django-timedeltafield>=0.6.7
south>=0.7.6
*.swp
